refactor(models): name files after primary resource
