--- src/models/moves/index.ts
@@ -1,1 +1,0 @@
-export * from "./moves";
