+++ src/models/pokemon/growth-rate.ts
