+++ src/models/pokemon/index.ts
@@ -1,10 +1,10 @@
 export * from "./ability";
-export * from "./characteristics";
+export * from "./characteristic";
 export * from "./egg-group";
 export * from "./gender";
-export * from "./growth-rates";
+export * from "./growth-rate";
 export * from "./nature";
 export * from "./pokeathlon-stat";
 export * from "./pokemon";
-export * from "./stats";
+export * from "./stat";
 export * from "./type";
